Ensure iac providers are in top-level node_modules
Importing providers dynamically will only work if the provider is adjacent or higher than the @notation/core. When testing apps within this multi-package repo, the @notation/core package symlinks to a location above the node_modules folder of the application package. This means that the call site of any dynamic imports within core do not resolve other node modules belonging to that application. To resolve this iac packages are symlinked into the root node_modules directory.