Allow API gateway route to be configured with a JWT authorizer (#20)
# Allow API gateway route to be configured with a JWT Authorizer

AWS API Gateway routes can be configured with authorizers, which control access to 
the route endpoints:

https://docs.aws.amazon.com/apigateway/latest/developerguide/http-api-access-control.html

JWT tokens (https://jwt.io/introduction) are commonly used for authorization to API endpoints
to prove the user is authorized to perform a given action. AWS provides the ability to
attach a 'JWT authorizer' to routes which will verify the token is valid and that the 
token claims contains at least one of the configured 'authorizationScopes' parameters for the endpoint
(https://datatracker.ietf.org/doc/html/rfc6749#section-3.3).

# Changes

## IAC layer

new 'RouteAuth' resource, capable of representing a route authorizer config (also can represent
custom lambda authorizers which are not yet supported, etc.) It requires a 'API gateway' as a dependency to
extract the API ID. A route requires this as an *optional* dependency (routes can be public.)

![dependency_graph](https://github.com/notation-dev/notation/assets/1145180/ab91d163-f175-461b-b2cb-bbd9dc2bc4e1)

## AWS API layer

route resource - can be configured with an authorizer. It sets up the necessary resource dependency wiring 
(as described above.) `packages/aws/src/api-gateway/auth.ts` has some types for representing the auth config
which will later be expanded for different authorizer types.

router resource - new `withJWTAuthorizer` function makes it possible to build a type safe route handler with a event handler that expects to find a `claims` object with the `<ClaimsType>` generic parameter. Example usage can be found in the `examples` folder.

`packages/aws/src/lambda.fn/handlers.ts` - new `jwtAuthorizedApiRequest` handler allows a handler to be passed in that expects a JWT token with the generic parameter's type fields on the `context.requestContext` field. This was achieved via the new `JWTAuthorizedApiGatewayHandler` type which overrides the AWS sdk with the `APIGatewayProxyEventV2WithJWTAuthorizer` type, replacing the `claims` type (see `packages/aws/src/shared/lambda.handler.ts)

## Orchestration Layer

* Optional dependencies can now be represented (via a nullable type.) This is used to support route's optional
authorizer dependency.


# Assumptions

We assume the client will be sending the JWT token in the 'Authorization' header with the prefix 'Bearer ' as
is standardised (https://www.rfc-editor.org/rfc/rfc6750.) 

AWS can be configured via the 'IdentitySource' param to look anywhere in the HTTP request for this token.
However, in this PR, we've hard coded this value to '$request.header.Authorization'.
(https://docs.aws.amazon.com/apigatewayv2/latest/api-reference/apis-apiid-authorizers-authorizerid.html#apis-apiid-authorizers-authorizerid-prop-authorizer-identitysource)

We've hardcoded this value as it follows the RFC and guides the user into following it.

# Future Work

* Custom Lambda authorizers
* IAM based authorization
* Iterating on the 'router' builder pattern to improve it